+++ volmdlr/faces.py
@@ -826,6 +826,8 @@ class Surface3D(DessiaObject):
                 outer_contour2d, inner_contours2d = self.repair_contours2d(contours2d[0], contours2d[1:])
             else:
                 for contour2d in contours2d:
+                    if not contour2d.is_ordered(1e-4):
+                        contour2d = vm_parametric.contour2d_healing(contour2d)
                     inner_contours2d.append(contour2d)
                     contour_area = contour2d.area()
                     if contour_area > area:
@@ -839,6 +841,8 @@ class Surface3D(DessiaObject):
             class_ = globals()[self.face_class]
         else:
             class_ = self.face_class
+        if not outer_contour2d.is_ordered(1e-4):
+            outer_contour2d = vm_parametric.contour2d_healing(outer_contour2d)
         surface2d = Surface2D(outer_contour=outer_contour2d,
                               inner_contours=inner_contours2d)
         return class_(self, surface2d=surface2d, name=name)
@@ -956,11 +960,11 @@ class Surface3D(DessiaObject):
                     if primitives is None:
                         continue
                     primitives3d.extend(primitives)
-                except NotImplementedError:
+                except AttributeError:
                     print(f'Class {self.__class__.__name__} does not implement {method_name}'
                           f'with {primitive2d.__class__.__name__}')
             else:
-                raise NotImplementedError(
+                raise AttributeError(
                     f'Class {self.__class__.__name__} does not implement {method_name}')
 
         return volmdlr.wires.Contour3D(primitives3d)
@@ -1885,6 +1889,8 @@ class PeriodicalSurface(Surface3D):
         theta1, z1 = linesegment2d.start
         theta2, z2 = linesegment2d.end
         if math.isclose(theta1, theta2, abs_tol=1e-4) or linesegment2d.name == "parametic.linesegment":
+            if self.point2d_to_3d(linesegment2d.start).is_close(self.point2d_to_3d(linesegment2d.end)):
+                print(True)
             return [vme.LineSegment3D(self.point2d_to_3d(linesegment2d.start),
                                       self.point2d_to_3d(linesegment2d.end))]
 
@@ -1943,7 +1949,7 @@ class CylindricalSurface3D(PeriodicalSurface):
             fig = plt.figure()
             ax = fig.add_subplot(111, projection='3d')
 
-        self.frame.plot(ax=ax)
+        self.frame.plot(ax=ax, ratio=0.5 * length)
         for i in range(nlines):
             theta = i / (nlines - 1) * volmdlr.TWO_PI
             start = self.point2d_to_3d(volmdlr.Point2D(theta, -0.5 * length))
@@ -2721,8 +2727,8 @@ class ConicalSurface3D(PeriodicalSurface):
             fig = plt.figure()
             ax = fig.add_subplot(111, projection='3d')
 
-        self.frame.plot(ax=ax)
-        x = z / math.sin(self.semi_angle)
+        self.frame.plot(ax=ax, ratio=z)
+        x = z * math.tan(self.semi_angle)
         # point1 = self.frame.local_to_global_coordinates(volmdlr.Point3D(-x, 0, -z))
         point1 = self.frame.origin
         point2 = self.frame.local_to_global_coordinates(volmdlr.Point3D(x, 0, z))
@@ -2850,7 +2856,8 @@ class ConicalSurface3D(PeriodicalSurface):
             start = volmdlr.Point2D(end.x, 0)
         elif start.x != end.x and end == volmdlr.Point2D(0, 0):
             end = volmdlr.Point2D(start.x, 0)
-
+        elif start.x != end.x:
+            end = volmdlr.Point2D(start.x, end.y)
         if not start.is_close(end):
             return [vme.LineSegment2D(start, end)]
         return [vme.BSplineCurve2D.from_points_interpolation([start, end], 1, False)]
@@ -3084,11 +3091,11 @@ class SphericalSurface3D(PeriodicalSurface):
                         primitives3d.extend(primitives_list)
                     else:
                         continue
-                except NotImplementedError:
+                except AttributeError:
                     print(f'Class {self.__class__.__name__} does not implement {method_name}'
                           f'with {primitive2d.__class__.__name__}')
             else:
-                raise NotImplementedError(f'Class {self.__class__.__name__} does not implement {method_name}')
+                raise AttributeError(f'Class {self.__class__.__name__} does not implement {method_name}')
 
         return volmdlr.wires.Contour3D(primitives3d)
 
@@ -4382,7 +4389,7 @@ class BSplineSurface3D(Surface3D):
         else:
             i = 1
         min_bound, max_bound = self.surface.domain[i]
-        delta = max_bound - min_bound
+        delta = max_bound + min_bound
 
         if math.isclose(start[i], min_bound, abs_tol=1e-4) and pt_after_start[i] > 0.5 * delta:
             start[i] = max_bound
@@ -9569,7 +9576,11 @@ class OpenShell3D(volmdlr.core.CompositePrimitive3D):
     def triangulation(self):
         meshes = []
         for face in self.faces:
-            face_mesh = face.triangulation()
+            try:
+                face_mesh = face.triangulation()
+            except Exception:
+                warnings.warn("Could not triangulate face. Probabaly because topology error in contour2d.")
+                continue
             meshes.append(face_mesh)
         return vmd.DisplayMesh3D.merge_meshes(meshes)